+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow arbitrary to be passed to RosterProvider
 - Allow for simlucast configuration in MeetingProvider
 - Updated button focus states to increase their accessibility
+- [Documentation] Update README with solutions of common issues
 
 ### Removed